netstacklat: Generalize initialization of array filter maps
The filtering for specific pids (--pids/-p) makes use of a BPF array
map where each entry to be included is set to 1 (the rest remain 0 as
all entries by default are zeroed in array maps). Generalize the user
space logic that initializes the entries in this filter map so that it
can be reused by other similar features in the future.

Furthermore, change the value type of the filter map from u8 to
u64. While only one bit is really required, each BPF map entry takes
up at least 8 bytes regardless, so using u8 does not save any
space. Additionally, a future commit will add functionality to also
filter by cgroup, and to keep that feature compatible with
ebpf-exporter the value size needs to be 8 bytes. So use u64 keys for
all filter maps from here on out to make it easier to reuse logic.

+++ netstacklat/netstacklat.bpf.c
@@ -41,7 +41,7 @@ struct {
 	__uint(type, BPF_MAP_TYPE_ARRAY);
 	__uint(max_entries, PID_MAX_LIMIT);
 	__type(key, u32);
-	__type(value, u8);
+	__type(value, u64);
 } netstack_pidfilter SEC(".maps");
 
 static u64 *lookup_or_zeroinit_histentry(void *map, const struct hist_key *key)
@@ -161,7 +161,7 @@ static void record_skb_latency(struct sk_buff *skb, enum netstacklat_hook hook)
 
 static bool filter_pid(u32 pid)
 {
-	u8 *pid_ok;
+	u64 *pid_ok;
 
 	if (!user_config.filter_pid)
 		// No PID filter - all PIDs ok
